refactor: split ClientPage into smaller components and add back button
